refactor(frame): remove pins data
