+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -10,10 +10,6 @@ import java.util.Objects;
 public class NormalFrame implements Frame {
     private static final int FIRST_FRAME = 1;
     private static final int FINAL_FRAME = 10;
-    private static final int PINS_LIMIT = 10;
-    private static final String PIN_MAX_ERROR = "핀의 합계가 10개보다 클 수 없습니다.";
-    private static final int CAN_ROLL_LIMIT = 2;
-    private static final int FIRST_ROLL = 1;
 
     private Pins pins;
     private int index;
@@ -23,47 +19,25 @@ public class NormalFrame implements Frame {
         this.index = index;
     }
 
-    public NormalFrame() {
-    }
-
     public static NormalFrame firstFrame() {
         return new NormalFrame(FIRST_FRAME);
     }
 
-    public static NormalFrame frame() {
-        return new NormalFrame();
-    }
-
     @Override
     public void roll(int pin) {
-        if (isPinTotalOverTen(pin)) {
-            throw new IllegalArgumentException(PIN_MAX_ERROR);
-        }
         if (canRoll()) {
-            pins.addPins(pin);
+            pins.addPins(this, pin);
         }
     }
 
     @Override
     public boolean canRoll() {
-        return !isAlreadyStrike() && !isRolledTwice();
+        return pins.isPinReady();
     }
 
     @Override
     public boolean isGameOver() {
         return index != FINAL_FRAME;
-    }
-
-    private boolean isRolledTwice() {
-        return pins.rollCount() >= CAN_ROLL_LIMIT;
-    }
-
-    private boolean isAlreadyStrike() {
-        return pins.rollCount() == FIRST_ROLL && pins.getTotalPins() == PINS_LIMIT;
-    }
-
-    private boolean isPinTotalOverTen(int pin) {
-        return this.getTotal() + pin > PINS_LIMIT;
     }
 
     @Override
